pyproject.toml
[project]
name = "fri-g2g-upload-tl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gspread>=6.2.0",
    "httpx>=0.28.1",
    "pydantic>=2.11.3",
    "pydoll-python>=1.7.0",
    "python-dotenv>=1.1.0",
]
